+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+## v1.2
+
+## v1.2.7
+- Experimental feature: real time draping of GeoJSON overlays (see example).
+- Add a `Viewport` structure required on initial load and updated via the `setViewport()` function. Contains `width`, `height`, and `devicePixelRatio`.
+- Added `setRenderer()` for setting the renderer after the first load if needed.
+
+## v1.2.3
+ - runtime.update() now takes `viewportSize:Vector2` rather than `viewportHeight:number` (for future features that will require viewport width).
+ - Preliminary support and example for loading GeoJSON and overlaying on Google 3D Tiles.
+
+## v1.2.2
+ - Extract data attributions from Google's 3D Tiles.
+ - Options added: 
+    - `collectAttributions` 
+    - `contentPostProcess`
+    - `resetTransform`
+ - Options removed:
+    - `geoTransform` (use `orientToGeocoord` and `getWebMercatorCoord`)
+    - `shaderCallback` (use `contentPostProcess`)
+    - `computeNormals` (use `contentPostProcess`)
+
+## v1.2.0
+ - Support for Photorealistic 3D Tiles by Google Maps ([demo: requires API key](https://nytimes.github.io/three-loader-3dtiles/dist/web/examples/demos/google-3dtiles/), [demo source](https://github.com/nytimes/three-loader-3dtiles/blob/main/examples/demos/google-3dtiles/index.html)).
+ - Upgrade to loaders.gl v4.1 and Three.js r160.
+   - Includes loaders.gl PRs [#2847](https://github.com/visgl/loaders.gl/pull/2847) and [#2851](https://github.com/visgl/loaders.gl/pull/2851).
+ - Faster build system with Vite.
+ - Support for `memoryAdjustedScreenSpaceError` and `memoryCacheOverflow` - Adjust the maximum screen space error based on GPU memory utilization.
+ - More accurate estimates of GPU memory utilization.
+ - `runtime.update()` function now takes `viewportHeight` instead of the whole `renderer` object (for screen space error calculations).
+ - Support providing an external `gltfLoader` in `options`, to reuse existing loader workers.
+ - Minor performance optimizations.
+
 ## v1.1
 
 ### v1.1.17